refactor: clarify remote runtime structure
